+++ demo/dashboard/client/src/pages/files/File.js
@@ -14,7 +14,6 @@ import FormUpload from '../../components/FormUpload';
 import LogsTab from './file/Logs';
 import ProjectsTab from './file/Projects';
 import ReportsTab from './file/Reports';
-import EnrichmentTab from './file/Enrichment';
 
 import './File.css';
 
@@ -240,7 +239,6 @@ class File extends React.Component {
           <span className="ecl-icon ecl-icon--left" />
           Go Back to My Files
         </Link>
-
         <div className="ecl-row ecl-u-mv-m">
           <div className="ecl-col-md-4 ecl-u-d-flex ecl-u-justify-content-center ecl-u-align-items-baseline">
             <span className="ecl-u-fs-xxl">
@@ -260,7 +258,6 @@ class File extends React.Component {
             </span>
           </div>
         </div>
-
         <nav className="ecl-navigation-list-wrapper">
           <h2 className="ecl-u-sr-only">Navigation Menu</h2>
           <ul className="ecl-navigation-list ecl-navigation-list--tabs">
@@ -294,15 +291,6 @@ class File extends React.Component {
             </li>
             <li className="ecl-navigation-list__item">
               <NavLink
-                to={`${match.url}/enrichment`}
-                className="ecl-navigation-list__link ecl-link"
-                activeClassName="ecl-navigation-list__link--active"
-              >
-                Enrichment
-              </NavLink>
-            </li>
-            <li className="ecl-navigation-list__item">
-              <NavLink
                 to={`${match.url}/reports`}
                 className="ecl-navigation-list__link ecl-link"
                 activeClassName="ecl-navigation-list__link--active"
@@ -312,7 +300,6 @@ class File extends React.Component {
             </li>
           </ul>
         </nav>
-
         <Route
           exact
           path={`${match.url}`}
@@ -350,10 +337,6 @@ class File extends React.Component {
         <Route
           path={`${match.url}/projects`}
           render={() => <ProjectsTab computedKey={computedKey} />}
-        />
-        <Route
-          path={`${match.url}/enrichment`}
-          render={() => <EnrichmentTab computedKey={computedKey} />}
         />
         <Route
           path={`${match.url}/reports`}
